alkahest: Pre-Processing XY Data from Experimental Methods

A lightweight, dependency-free toolbox for pre-processing XY data from experimental methods (i.e. any signal that can be measured along a continuous variable). This package provides methods for baseline estimation and correction, smoothing, normalization, integration and peaks detection. Baseline correction methods includes polynomial fitting as described in Lieber and Mahadevan-Jansen (2003) doi:10.1366/000370203322554518 , Rolling Ball algorithm after Kneen and Annegarn (1996) doi:10.1016/0168-583X(95)00908-6 , SNIP algorithm after Ryan et al. (1988) doi:10.1016/0168-583X(88)90063-8 , 4S Peak Filling after Liland (2015) doi:10.1016/j.mex.2015.02.009 and more.

Details

Package:	alkahest

Type:	Package

Version:	1.1.1

License:	GPL-3

Zenodo:	doi:10.5281/zenodo.7081524

See also

Author

Full list of authors and contributors (alphabetic order)

Nicolas Frerebeau	Université Bordeaux Montaigne, France

Brice Lebrun	Université Bordeaux Montaigne, France

Package maintainer

Nicolas Frerebeau
nicolas.frerebeau@u-bordeaux-montaigne.fr

Archéosciences Bordeaux (UMR 6034)
Maison de l'Archéologie
Université Bordeaux Montaigne
33607 Pessac cedex
France